refactor part.c -> rom.c
