+++ src/main.c
@@ -3,9 +3,9 @@
 #include <string.h>
 #include <unistd.h>
 
-#include "arc.h"
 #include "mra.h"
-#include "part.h"
+#include "rom.h"
+#include "arc.h"
 #include "utils.h"
 
 // make vscode happy
